implementa testes

+++ src/main/java/org/bridge/Motor.java
@@ -2,5 +2,5 @@ package org.bridge;
 
 public interface Motor {
 
-    float custoConsumo();
+    double custoConsumo();
 }
